requirements.txt
awscliv2==1.5.2
boto3==1.20.21
boto3-stubs==1.20.21
botocore==1.23.21
botocore-stubs==1.23.21
jmespath==0.10.0
mypy-boto3-cloudwatch==1.20.9
mypy-boto3-lambda==1.20.13
mypy-boto3-rds==1.20.12
mypy-boto3-s3==1.20.17
mysql-connector-python==8.0.27
protobuf==3.19.1
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
